backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
celery>=5.3
redis>=5.0
psycopg2-binary>=2.9
gunicorn>=21.0
python-dotenv>=1.0
openai>=1.0
gtts>=2.5
Pillow>=10.0
boto3>=1.34
django-storages>=1.14
dj-database-url>=2.1
